fix: fuser used-before-assignment, token re-encoding, deduplicate model loading
- Fix UnboundLocalError: construct Fuser before perplexity and eval_text
  branches so --eval --model fused works without --perplexity
- Replace hrm_to_ouro direct token mapping with ouro_tok.encode(generated_text)
  to prevent the "the" → t,h,e fragmentation described in AGENTS.md
- Consolidate three duplicate model-loading implementations into loader.load_all()
  returning LoadedModels dataclass (models + tokenizers + matcher + device)
- Document all 15 fusion strategies (was 11) in README, CLI help, and bibtex
- Add explicit dtype=torch.long to all torch.tensor() calls for input_ids
  to prevent FloatTensor when passing Python lists to embedding layers

+++ src/llm_fusion/metrics.py
@@ -226,10 +226,10 @@ def evaluate_text(
 
         with torch.no_grad():
             ouro_out = ouro_model(
-                torch.tensor([ouro_prefix_ids], device=device),
+                torch.tensor([ouro_prefix_ids], device=device, dtype=torch.long),
             )
             hrm_out = hrm_model(
-                torch.tensor([seq[:t]], device=device),
+                torch.tensor([seq[:t]], device=device, dtype=torch.long),
                 token_type_ids=torch.ones(t, dtype=torch.long, device=device).unsqueeze(0),
             )
